Updated files with codes
Updated the files with codes and filtered for the counties in our data set. There are only 10 counties missing from this data set. One main issue is that the census and ACS data looks like it stops at 2015. We may have trouble seeing results for the ones that were opened close to 2015 or after.